+++ Cartao/Product/CartaoCredito.cs
@@ -3,7 +3,7 @@
     public abstract class CartaoCredito
     {
         public abstract string Tipo { get; }
-        public abstract float Limite { get; set; }
-        public abstract float CobrancaAnual { get; set; }
+        public abstract double Limite { get; set; }
+        public abstract double CobrancaAnual { get; set; }
     }
 }
