+++ Cartao/Product/EntregaProduct.cs
@@ -4,6 +4,6 @@
     {
         public abstract string TipoEntrega { get; }
         public abstract int Prazo { get; set; }
-        public abstract int Valor { get; set; }
+        public abstract double Valor { get; set; }
     }
 }
